Added logos | refactoring
