+++ utils/skillsData.ts
@@ -1,0 +1,146 @@
+import nextjs from "../public/imgs/skills/nextjs.svg";
+import react from "../public/imgs/skills/react.svg";
+import express from "../public/imgs/skills/express.svg";
+import tensorflow from "../public/imgs/skills/tensorflow.svg";
+import tailwindcss from "../public/imgs/skills/tailwindcss.svg";
+import typescript from "../public/imgs/skills/typescript.svg";
+import javascript from "../public/imgs/skills/javascript.svg";
+import python from "../public/imgs/skills/python.svg";
+import cpp from "../public/imgs/skills/cplusplus.svg";
+import postgresql from "../public/imgs/skills/postgresql.svg";
+import mongodb from "../public/imgs/skills/mongo.svg";
+import mysql from "../public/imgs/skills/mysql.svg";
+import azure from "../public/imgs/skills/azure.svg";
+import aws from "../public/imgs/skills/aws.svg";
+import appwrite from "../public/imgs/skills/appwrite.svg";
+import firebase from "../public/imgs/skills/firebase.svg";
+import supabase from "../public/imgs/skills/supabase.svg";
+import kubernetes from "../public/imgs/skills/kubernetes.svg";
+import docker from "../public/imgs/skills/docker.svg";
+import gitlab from "../public/imgs/skills/gitlab.svg";
+
+
+const skillsData = [
+  {
+    name: "React.js",
+    type: "frameworks",
+    light: react,
+    dark: "path/to/dark/react-icon.png",
+  },
+  {
+    name: "Next.js",
+    type: "frameworks",
+    light: nextjs,
+    dark: "/public/imgs/skills/nextjs.svg",
+  },
+  {
+    name: "Express",
+    type: "frameworks",
+    light: express,
+    dark: "path/to/dark/express-icon.png",
+  },
+  {
+    name: "TensorFlow",
+    type: "frameworks",
+    light: tensorflow,
+    dark: "path/to/dark/tensorflow-icon.png",
+  },
+  {
+    name: "Tailwind CSS",
+    type: "frameworks",
+    light: tailwindcss,
+    dark: "path/to/dark/tailwindcss-icon.png",
+  },
+  {
+    name: "TypeScript",
+    type: "Languages",
+    light: typescript,
+    dark: "path/to/dark/typescript-icon.png",
+  },
+  {
+    name: "JavaScript",
+    type: "Languages",
+    light: javascript,
+    dark: "path/to/dark/javascript-icon.png",
+  },
+  {
+    name: "Python",
+    type: "Languages",
+    light: python,
+    dark: "path/to/dark/python-icon.png",
+  },
+  {
+    name: "C++/C",
+    type: "Languages",
+    light: cpp,
+    dark: "path/to/dark/cpp-icon.png",
+  },
+  {
+    name: "PostgreSQL",
+    type: "Databases",
+    light: postgresql,
+    dark: "path/to/dark/postgresql-icon.png",
+  },
+  {
+    name: "MongoDB",
+    type: "Databases",
+    light: mongodb,
+    dark: "path/to/dark/mongodb-icon.png",
+  },
+  {
+    name: "MySQL",
+    type: "Databases",
+    light: mysql,
+    dark: "path/to/dark/mysql-icon.png",
+  },
+  {
+    name: "Azure",
+    type: "Cloud Services and DevOps",
+    light: azure,
+    dark: "path/to/dark/azure-icon.png",
+  },
+  {
+    name: "AWS",
+    type: "Cloud Services and DevOps",
+    light: aws,
+    dark: "path/to/dark/aws-icon.png",
+  },
+  {
+    name: "Appwrite",
+    type: "Cloud Services and DevOps",
+    light: appwrite,
+    dark: "path/to/dark/appwrite-icon.png",
+  },
+  {
+    name: "Firebase",
+    type: "Cloud Services and DevOps",
+    light: firebase,
+    dark: "path/to/dark/firebase-icon.png",
+  },
+  {
+    name: "Supabase",
+    type: "Cloud Services and DevOps",
+    light: supabase,
+    dark: "path/to/dark/supabase-icon.png",
+  },
+  {
+    name: "Kubernetes",
+    type: "Cloud Services and DevOps",
+    light: kubernetes,
+    dark: "path/to/dark/kubernetes-icon.png",
+  },
+  {
+    name: "Docker",
+    type: "Cloud Services and DevOps",
+    light: docker,
+    dark: "path/to/dark/docker-icon.png",
+  },
+  {
+    name: "GitLab",
+    type: "Cloud Services and DevOps",
+    light: gitlab,
+    dark: "path/to/dark/gitlab-icon.png",
+  },
+];
+
+export default skillsData;
